[COPOR-2] 패키지 정리
