--- copang-order-api/src/main/kotlin/com/copang/product/ProductsHttpRequest.kt
@@ -1,5 +1,0 @@
-package com.copang.product
-
-data class ProductsHttpRequest(
-    val productIds: List<Long>,
-)
