+++ copang-order-api/src/main/kotlin/com/copang/repository/product/ProductsHttpRequest.kt
@@ -1,0 +1,5 @@
+package com.copang.repository.product
+
+data class ProductsHttpRequest(
+    val productIds: List<Long>,
+)
