+++ copang-order-application/src/test/kotlin/com/copang/cart/CartServiceTest.kt
@@ -1,13 +1,13 @@
 package com.copang.cart
 
-import com.copang.exception.CopangException
-import com.copang.exception.ErrorType
+import com.copang.common.exception.CopangException
+import com.copang.common.exception.ErrorType
 import com.copang.mock.CartTestData
 import com.copang.mock.ProductTestData
 import com.copang.mock.UserInfoTestData
 import com.copang.product.Product
 import com.copang.product.ProductRepository
-import com.copang.utils.AuthUtils
+import com.copang.common.utils.AuthUtils
 import io.kotest.assertions.throwables.shouldThrowExactly
 import io.kotest.matchers.shouldBe
 import io.mockk.every
