+++ copang-order-domain/src/main/kotlin/com/copang/common/constants/DateConstants.kt
@@ -1,0 +1,7 @@
+package com.copang.common.constants
+
+import java.time.ZoneId
+
+object DateConstants {
+    val ZONE_KST: ZoneId = ZoneId.of("Asia/Seoul")
+}
