+++ copang-order-domain/src/main/kotlin/com/copang/common/constants/TypeConstants.kt
@@ -1,0 +1,7 @@
+package com.copang.common.constants
+
+object TypeConstants
+
+const val EMPTY_STRING = ""
+const val EMPTY_ID = 0L
+const val EMPTY_NUMBER = 0
