--- copang-order-domain/src/main/kotlin/com/copang/constants/DateConstants.kt
@@ -1,7 +1,0 @@
-package com.copang.constants
-
-import java.time.ZoneId
-
-object DateConstants {
-    val ZONE_KST: ZoneId = ZoneId.of("Asia/Seoul")
-}
