--- copang-order-domain/src/main/kotlin/com/copang/constants/TypeConstants.kt
@@ -1,7 +1,0 @@
-package com.copang.constants
-
-object TypeConstants
-
-const val EMPTY_STRING = ""
-const val EMPTY_ID = 0L
-const val EMPTY_NUMBER = 0
