--- copang-order-domain/src/main/kotlin/com/copang/exception/CopangException.kt
@@ -1,29 +1,0 @@
-package com.copang.exception
-
-import java.lang.RuntimeException
-
-class CopangException(
-    message: String = "",
-    val errorCode: String = "",
-    cause: Throwable? = null,
-) : RuntimeException(message, cause) {
-
-    constructor(
-        errorType: ErrorType,
-        cause: Throwable? = null,
-    ) : this(
-        message = errorType.message,
-        errorCode = errorType.errorCode,
-        cause = cause,
-    )
-
-    constructor(
-        errorType: ErrorType,
-        cause: Throwable?,
-        message: String = "",
-    ) : this(
-        message = message.ifBlank { errorType.message },
-        errorCode = errorType.errorCode,
-        cause = cause,
-    )
-}
