+++ copang-order-infrastructure/build.gradle.kts
@@ -12,7 +12,6 @@ dependencies {
     implementation("org.springframework.boot:spring-boot-starter-data-jpa")
     runtimeOnly("org.postgresql:postgresql")
     runtimeOnly("com.h2database:h2")
-    runtimeOnly("com.fasterxml.jackson.module:jackson-module-kotlin")
 
     testImplementation(project(":tests"))
 }
