--- copang-order-infrastructure/src/main/kotlin/com/copang/JpaConfig.kt
@@ -1,8 +1,0 @@
-package com.copang
-
-import org.springframework.context.annotation.Configuration
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories
-
-@Configuration
-@EnableJpaRepositories(basePackages = ["com.copang"])
-class JpaConfig
