+++ copang-order-infrastructure/src/main/kotlin/com/copang/common/config/JpaConfig.kt
@@ -1,0 +1,8 @@
+package com.copang.common.config
+
+import org.springframework.context.annotation.Configuration
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories
+
+@Configuration
+@EnableJpaRepositories(basePackages = ["com.copang"])
+class JpaConfig
